Big refactoring - Interim Commit
The old LFE based compile chain has been taken out in favour of Erlang/Merl.

LFE has been ditched as a dependency.

The rational behind this is that there will need to be a source map from the Pometo source to the compiled source to enable proper error handling.

The parser needs to be readjusted. In the interpreter the parsed code has the variable declarations substituted out for their current values - in the compiler those variable names are captured and managed by the Erlang compiler/runtime.

To that end the function that descends the AST tree now either subsitutes or validates that the binding is already declared depending on if you are interpreting or compiling.

The beam files are compiling and loading - but the exported functions have no body and merely return the string "randomette" at the moment.