Final commit of new compiler/interpreter infrastructure.
This is a major refactor which substantially unifies handling of error messages:

* in the lexer
* in the parser
* validating that invoked variables are bound (for the interpreter/rappel)
* in the compilation of the Erlang code
* at runtime

The generated Erlang modules will have source code mapping on run-time errors as explained in the Design Documents section ***Source Code Mapping***

The final implemenation of the source mapping (error trapping and correct line number handling for the generated source code (the function `reset/2` in `pometo_compiler.erl`) will not be fixed up until after the `lexer`/`parser` can generate externally exported functions.

`Pometo` had been expanded to recognise full unicode in variables this has been rolled back to simplify the handling of variable names (ie `Pometo` variable names now match `Erlang` ones. This might be reversed in the future when the compiler is more stable.

Unicode names will need to be converted to lists, hashed and base16 encoded. This isn't difficult but it would be nice to be able to annotate the source code with the mappings, and TBQH I can't be arsed figuring it out the noo ;-)

In addition bugs arising from trying to execute lines consisting only of comments in both the interpreted and compiled paths have been fixed.

The use of 'let' an Erlang reserved word in the `pometo_runtime` was a right pain in construction of compilable modules so it has been transformed out into the straightforward variable `runtime_let`.

The test helper has been standardised (with appropriate changes to the `pometo_docs_to_tests` rebar plugin).

Additional compiler/compiler failure tests have been worked up.

The interpreter and compiler tests generated from the docs now only both return the results of the last executed statement.