Big mix of changes.
At the core is a big implementation of complex numbers:

* the basis of scalar extensions
* how we can extend primitives to tupes, records and maps

As part of this there is a host of other minor but important work:

* rename of liffey record to ast as we are not using LFE anymore ;-)
* short form macros to make code more readable
* raw expressions are considered acceptable - they return themselves.

The main non-functional thing is work on making a better developer experience including some pages for developers and how to get started.

Part of this is more test shadowing and a description of the shadow structure that allows new developers to break loads of tests and be confident that they can fix me.

Structuring tests around shadowing lets the test suite tell you which broken tests to fix first when you get massive test failures and is critical to a happy developer experience.

Previously we had the core tests in `test/` which test the basics of the lexer/parser and then the bulk of the tests in 'test/generated_tests` from the documentation.

The generated tests are ***shadowed*** by the basic tests - if the basic tests fail then sure as hell the feature tests will fail.

We now add a new set of shadowing tests - not as primitive as those in `test/`. These basic tests are on the page `docs/implementation_reference/basics.md`. They check that the core runtime custom paths are correctly executed for each branch.