Big final commit of proper handling of arrays.
Summary of changes:

There are not clear differences between arrays and scalars (it is possible to have a scalar of an array as well).

The shape operator rho works.

There is a detailed formatted that follows the Dyalog one is pretty much most details (some issues around how and when we truncate data - and double bumping in matrices with more that 3 dimensions).

In addition there is now a smoke test suite and a batch job that runs the non-generated-tests in the appropriate shadowed order (ie the order in which you should fix the failing tests).

A sketch of iota has been included in the documentation but iota has not been implemented in the runtime so the tests are commented out (there is also a rho test in that state - it is running ahead of the compiler).

A new special hand written test suite to test formatting is included - if the formatted is borked loads of tests will go down (obvs).