Range of stuff
First up is helper functions to help develop the different internal representation of shapes in the runtime.

These are explained in the docs page stdlib_reference/make_indexed_function

The whole jekyll stuff has been adjusted too - so that there is a consistent logo, a way to click back to docs home and stuff.

Various beginnings of the underpinnings of automagic dynamic indexing and lazy/eager conversion being put in place.