This commit puts in place indexed vectors and handles conversion from unindexed to indexed (and back again).
Part of that is building out standard functions that operate on arguments with one case for lists and another for maps.

`pometo_docs_to_tests` has been updated to generate a 4th set of tests from the documentation leading to a big uptick of tests fo 459.

Quite a big set of changes for such a short commit msg ;-)