The major part of this commit is the adding the building of trains, fgh forks, Agh forks and atop (dyadic and monadic). This is not working yet, but the infrastructure is now passing with the old code base. (The way in which we build dyadic/monadic functions has changed).
Big changes to the runtime. Array types now include `func` which are arrays of functions and `maybe_func` - the contain a mixture of `#$func¯{}` records and `#'$var¯'{}`s which might be either data or functions.

The clean train path (ie for `func` types) is working - but the resolution of `maybe_func` is not yet implemented.

Implementing `func` and `maybe_func` arrays involved threading in defered evaluation.

* bug fixes for the return types (eg 1 + 1 where both 1's are scalars should return a scalar - but if one is cast to a vector of length 1 then the return type is an array)
  * and associated documentation/tests to that effect in the `functions` section
* minor utility fn in a header file for errors
* a scope value in the process dictionary for declaring variables inside a scope to stop scope leaking out
* changed the name of the `do` from `let` to `let_op` becuz Erlang reserved word - easier and cleaner than custom quoting of atoms
* changes to the ordered test batch job to make it cleaner
* tidy up of the documentation
	* unlabled text blocks
	* typos

The handling of 0 in the `/` and `⌿` operators if borked (as is ¯ve LHS values) and this is mentioned in the points for improvement log. Also a path problem with the `Pometo` logo when running `Jekyll` locally.