Major rejigging of compiler internals
In particular the resultion of ambiguous results.

Most of this is described in the page ***parser notes*** which has details of the decision tree in the parser.

Trains are still not working but the base work for them is now in place.

Major changes in the internal data representation - particularly accumulation of vectors from a list of scalars.

New tests for making vectors from lists of shapes - added to `ordered_test.sh` job.

Documentation of debugging and working with parser issues (every day is a school day for me!) and skeleton options for debugging commented out in `rebar.config`.

Couple of other wierd bugs addressed, in particular array collapse - so an array of the form:

```erlang
				 args = [1]}
```

At the mo we just collapse this for monadic operations but for rank addressing and stuff it probably needs to be preserved - gonnae be a shed load of corner cases.